servers/cc/requirements.txt
aiohttp==3.5.4
appdirs==1.4.3
arrow==0.12.1
astroid==2.1.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
beautifulsoup4==4.6.3
cached-property==1.5.1
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
defusedxml==0.6.0
Django==2.2
django-admin-rangefilter==0.5.1
django-allauth==0.40.0
django-cors-headers==2.4.0
django-dynamic-fixture==2.0.0
django-extensions==2.1.2
django-filter==2.0.0
django-jsonfield==1.0.1
django-modelcluster==4.3
django-password-reset==2.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-taggit==0.23.0
django-templated-mail==1.1.1
django-treebeard==4.3
django-url-filter==0.3.8
djangorestframework==3.8.2
djoser==2.0.3
draftjs-exporter==2.1.7
drf-extensions==0.3.1
enum-compat==0.0.2
Flask==1.0.2
Flask-Login==0.4.1
flask-mongoengine==0.9.5
Flask-Script==2.0.6
Flask-WTF==0.14.2
furl==1.2.1
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
gunicorn==19.9.0
hashID==3.1.4
html5lib==0.999999999
httpie==1.0.2
httplib2==0.11.3
idna==2.7
idna-ssl==1.1.0
importlib-metadata==0.19
isort==4.3.4
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
logmatic-python==0.1.7
lxml==4.4.1
mako==1.1.0
MarkupSafe==1.0
mccabe==0.6.1
mongoengine==0.15.0
more-itertools==7.2.0
multidict==4.5.2
oauth2client==4.1.3
oauthlib==3.1.0
openapi-codec==1.3.2
orderedmultidict==1.0
packaging==19.1
phonenumbers==8.10.2
Pillow==5.4.1
pluggy==0.12.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.8.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.5.0
pyee==6.0.0
Pygments==2.4.2
pylint==2.2.2
pymongo==3.7.2
pyparsing==2.4.2
pyppeteer==0.0.25
pysocks==1.7.0
pytest==5.0.1
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-mock==1.10.4
python-dateutil==2.7.3
python-json-logger==0.1.9
python3-openid==3.1.0
pytz==2018.5
querystring-parser==1.2.3
requests==2.19.1
requests-oauthlib==1.3.0
responses==0.10.6
rest-condition==1.0.3
rsa==4.0
simplejson==3.16.0
six==1.11.0
slackbot==0.5.4
slackclient==2.1.0
slacker==0.13.0
slackeventsapi==2.1.0
sqlparse==0.3.0
tld==0.9.3
tqdm==4.37.0
typed-ast==1.3.1
typing-extensions==3.7.4
Unidecode==1.0.22
untangle==1.1.1
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.7.2
virtualenvwrapper-win==1.2.5
wagtail==2.4
Wand==0.5.2
wapiti3==3.0.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.44.0
websockets==6.0
Werkzeug==0.14.1
Willow==1.1
wrapt==1.11.1
WTForms==2.2.1
XlsxWriter==1.1.8
yarl==1.3.0
yaswfp==0.9.3
zipp==0.5.2
